emails/requirements.txt
Flask==2.3.6
python-dotenv==1.0.1
openai==1.32.0
PyPDF2==3.1.0
pdfplumber==0.10.1
Werkzeug==2.3.7
